added tests for stripe hooks and moved test data in separate file
