+++ package.json
@@ -53,6 +53,7 @@
     "reflect-metadata": "0.1.13",
     "rxjs": "7.5.5",
     "stripe": "^9.14.0",
+    "supertest": "^6.2.4",
     "swagger-ui-express": "^4.4.0",
     "tslib": "2.4.0"
   },
